split hekit to sperate command modules
